animationイベント発火場所を指定
