+++ app/assets/stylesheets/chats.scss
@@ -183,8 +183,9 @@ h2{
 }
 
 .amigo-space{
-  background-color: black;
-  height: 100px;
+  background-color: rgba(0, 0, 0, 0);
+  width: 100px;
+  height: 40px;
 }
 
 .amigoicon{
